// src/types/images/ImageNodeGeneratedData.ts
export type ImageNodeGeneratedData = {
    name: string
    description?: string
    creator: string
    license: string
    tags?: string
    source: string
    image_url_original: string
    image_url_xxl?: string
    image_url_xl?: string
    image_url_l?: string
    image_url_m?: string
    image_url_s?: string
    image_url_xs?: string
}